Split events.rs into events/ directory module for maintainability
The 4392-line monolithic events.rs was difficult to navigate. Split it
into focused submodules with clear responsibilities:

- mod.rs: SystingProbeRecorder struct, trait impl, core event handlers
- probe.rs: Probe type definitions, parsing, and Display impls
- config.rs: JSON config structs and config loading methods
- writer.rs: Parquet and Perfetto output writing
- tests.rs: All existing tests (unchanged logic)

No behavioral changes — purely structural refactor.
